refactor(lns-cli): move sandbox args (SandboxArgs/SandboxCommand/…) into the sandbox module
